added in debug info and changed -e to --env
was running into an issue with handbrake having lines in the json file with multiple items separated with spaces in between. this would cause issues because it would look for the image to pull at the end instead of the next -e event. it's now doing single quotes around the env variables so it will keep the information combined.